feat: Add Kind cluster deployment Makefile and update local PVC path
This commit introduces a new `Makefile` to streamline the deployment and cleanup of a Kind cluster.

The Makefile includes `deploy-kind` and `clean-kind` targets, which generate a temporary Kind cluster configuration
with a dynamic `hostPath` based on the current working directory.

Additionally, the `local-pvc.yaml` file has been updated to correct the `hostPath` for the `airflow-dags` PVC,
changing it from `/mnt/worker/dags` to `/mnt/worker/dags/`.